fwdwlv 180

+++ laravel/fwdwlv/app/Http/Controllers/QuestionsController.php
@@ -69,7 +69,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-
+        if(\Gate::denies('uodate-question', $question))
+        {
+            abort(403, "Access Denied");
+        }
         return view('questions.edit',compact('question'));    
     }
 
@@ -82,6 +85,10 @@ class QuestionsController extends Controller
      */
     public function update(Request $request, Question $question)
     {
+        if(\Gate::denies('uodate-question', $question))
+        {
+            abort(403, "Access Denied");
+        }
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', 'Question Updated');
     }
@@ -94,6 +101,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question))
+        {
+            abort(403, "Access Denied");
+        }
         $question->delete();
         return redirect('/questions')->with('sucess', "Question Deleted");
     }
